requirements.txt
fastapi
uvicorn[standard]
jinja2
httpx
python-dotenv